refactor: 예약에 대한 검증 책임을 적절하게 분배
